Make This Repo Work As Of October 2017
This commit's main purpose is to modify the mxnet-lambda code so that
it works as of October 2017. This means pulling the resnet weights
from a currently available location, removing the .pyc files that were
causing .asnumpy() to silently fail, and calling MXNet in the correct way so that the
model doesn't crash with the issue discussed and solved here:
https://stackoverflow.com/questions/44947104/mxnet-label-shapes-dont-match-names-specified-by-label-names

This commit also adds in the config files necessary to run this function
locally using AWS SAM Local.